chore: use Transfer type, move to interfaces
